refactor: move carbon estimator component to components
