+++ src/app/components/carbon-estimation-table/carbon-estimation-table.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CarbonEstimationTableComponent, TableItem } from './carbon-estimation-table.component';
 import { CarbonEstimationUtilService } from '../../services/carbon-estimation-util.service';
 import { CarbonEstimation } from '../../types/carbon-estimator';
-import { EmissionsLabels } from '../../carbon-estimation/carbon-estimation.constants';
+import { EmissionsLabels } from '../carbon-estimation/carbon-estimation.constants';
 
 describe('CarbonEstimationTableComponent', () => {
   let component: CarbonEstimationTableComponent;
