+++ src/app/tech-carbon-estimator/tech-carbon-estimator.component.ts
@@ -10,7 +10,7 @@ import {
   inject,
 } from '@angular/core';
 import { CarbonEstimatorFormComponent } from '../components/carbon-estimator-form/carbon-estimator-form.component';
-import { CarbonEstimationComponent } from '../carbon-estimation/carbon-estimation.component';
+import { CarbonEstimationComponent } from '../components/carbon-estimation/carbon-estimation.component';
 import { CarbonEstimation, EstimatorValues } from '../types/carbon-estimator';
 import { CarbonEstimationService } from '../services/carbon-estimation.service';
 import { FormsModule } from '@angular/forms';
